// src/app/types/user/user.document.ts
/* ••[1]••••••••••••••••••••••••• user.document.ts •••••••••••••••••••••••••••••• */

export type UserDocument = {
    email: string;
    firstName: string;
    id: string;
    lastName: string;
    lastUpdated: number;
    lastUpdatedBy: string;
}